refactoring folders
prepare to VHRoot

appending environment
