+++ mods/widgets/audio/WidgetMP3Frames.py
