+++ mods/widgets/periph/WidgetSerialPort.py
